backend/requirements.txt
# FastAPI & Server
fastapi==0.128.8
uvicorn[standard]==0.27.0
pydantic==2.12.5
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
pyarrow>=15.0.0

# ML & Forecasting
prophet==1.3.0
scikit-learn==1.5.0
statsmodels==0.14.1
xgboost==3.2.0

# Data Sources
requests==2.33.0
tenacity==8.2.3

# Security
python-jose[cryptography]>=3.3.0
bcrypt<6.0.0
python-multipart>=0.0.9

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.24.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.2.2
httpx==0.28.1
